// src/api/products/types.d.ts
export interface createProductInterface {
    sku: string;
    productType: string;
    productName: string;
    productDescription: string;
    productPrice:string;
    productOldPrice: string;
    productStock:string;
    productRate: string;
    reviewQuantity: string;
    soldQuantity: string;
    productViews: string;
    sizeOne: string;
    sizeTwo: string;
    sizeThree: string;
    sizeFour: string;
    sizeFive: string;
    productImage: string;
    productThumbnail: string;
    categoryOne: string;
    categoryTwo: string;
    categoryThree: string;
    categoryFour: string;
    categoryFive: string;
    interestForYou: string;
    productProfit: string;
}